requirements.txt
flask
flask-cors
PyPDF2
langchain
google-generativeai
faiss-cpu
python-dotenv
langchain-google-genai
pillow